fab: Move common functionality to common.py
